Do things on non-Windows OSes, and change the locale encoding (#3)
* Attempt to adjust TextEncoding on non-Windows

If someone expects their code page to be 65001 on Windows, it seems
sensible that the TextEncoding should be utf8 on non-Windows
(especially since folks are asking for this behavior in
bos/criterion#205).

Unfortunately, this does require some breaking changes, since
`withCodePage` now can error at runtime on non-Windows if supplied
an unknown `CodePage`, and `withCodePageVerbosity` had to be renamed
to `withCodePageOptions`. Luckily, I'm not aware of anyone who
actually uses either function, so I think this breakage shouldn't be
noticeable.

Along the way, I did some much-needed refactoring, including:

* The introduction of a `System.IO.CodePage.Internal`, modules, which
  now houses the `Options` data type, among other things.
* Added support for the 1252 (Latin1) code page.

* Fix infinitely-looping thinko

* Also change the locale encoding

Changing the encoding for just stdin/stdout/stderr apparently isn't
enough for some applications. A notable example of an application
that requires the locale encoding to be changed is `bench`. See
bos/criterion#205 for more details.

* setLocaleEncoding is only offered on GHC 7.4+